swan utils to swan functions
